requirements.txt
Faker==4.0.2
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
python-dateutil==2.8.1
six==1.14.0
text-unidecode==1.3
tqdm==4.45.0
